Update Trigger
The trigger first checks if there is sufficient stock for the bookid passed in the insert (by checking StockList’s quantityinstock for the book). If there is sufficient stock, the order is inserted, and the StockList quantityinstock record for the book is decremented by the quantity ordered. The order is marked fulfilled by inserting 1 into the OrderList’s fulfilled attribute for the order. 

If there is insufficient stock, then the order is inserted into OrderList, but fulfilled is set to 0. 
A record of the order is added to the BackOrderList to aid with stock resupply.